+++ src/loginServices.js
@@ -878,7 +878,7 @@ export function checkOkapiSession(okapiUrl, store, tenant) {
  * @returns {Promise}
  */
 export function requestLogin(okapiUrl, store, tenant, data) {
-  const loginPath = config.useSecureTokens ? 'login-with-expiry' : 'login';
+  const loginPath = 'login-with-expiry';
   return fetch(`${okapiUrl}/bl-users/${loginPath}?expandPermissions=true&fullPermissions=true`, {
     body: JSON.stringify(data),
     credentials: 'include',