+++ index.js
@@ -44,6 +44,7 @@ export { getUserTenantsPermissions } from './src/queries';
 /* Hooks */
 export { useUserTenantPermissions } from './src/hooks';
 export { useModuleInfo } from './src/hooks';
+export { usePreferences } from './src/hooks';
 
 /* misc */
 export { supportedLocales } from './src/loginServices';