+++ src/components/MainNav/MainNav.js
@@ -69,7 +69,7 @@ const MainNav = () => {
     return () => {
       _unsubscribe();
     };
-  }, []); // eslint-disable-line
+  }, [location]); // eslint-disable-line
 
   // if the location changes, we need to update the current module/query resource.
   // This logic changes the visible current app at the starting side of the Main Navigation.